🎨 Connect with airdrop API and redirect page
