+++ components/TokenBar.vue
@@ -29,9 +29,4 @@
     />
   </div>
 </template>
-<script lang="ts">
-import { Vue, Component } from 'vue-property-decorator'
 
-@Component
-export default class TokenBar extends Vue {}
-</script>
