+++ locales/en.json
@@ -1,38 +1,39 @@
 {
+  "AirDrop.button.claim":"Claim airdrop",
+  "AirDrop.button.enter":"Enter",
+  "AirDrop.button.notify":"Get notify",
+  "AirDrop.button":"Airdrop",
+  "AirDrop.content.prelaunch":"LikeCoin airdrop are on vote. The quick, brown fox jumps over a lazy dog. DJs flock by when MTV ax quiz prog. Junk MTV quiz",
+  "AirDrop.content.subscription":"Thanks for your support on the cosmos ecosystem, there are $LIKE allocated to you inside the upcoming LikeCoin airdrop. Do not miss the airdrop, subscribe to the news letters, we will notify you once the airdrop is ready to claim",
+  "AirDrop.guide.connect.keplr":"Connect cosmos/osmosis wallet with Keplr broswer extension",
+  "AirDrop.guide.connect.LikerId":"Connect to LikerID with the Liker Land app",
+  "AirDrop.guide.connect.wallet":"Connect or insert below your address and verify how many $LIKE have been allocated to you inside the upcoming LikeCoin airdrop",
+  "AirDrop.label.$like":"$LIKE",
+  "AirDrop.label.amount.airdrop":"Airdrop Amount",
+  "AirDrop.label.amount.umclaimed":"Unclaimed Airdrop",
+  "AirDrop.label.and":"and",
+  "AirDrop.label.connect.address":"or insert your address",
+  "AirDrop.label.connect.wallet":"Connect your wallet",
+  "AirDrop.label.cosmos":"Cosmos",
+  "AirDrop.label.decay.start":"Time to Airdrop Decay start",
+  "AirDrop.label.decay":"Current Decay Factor",
+  "AirDrop.label.follow.LikeCoin":"Follow LikeCoin",
+  "AirDrop.label.follow.Twitter":"Follow us on Twitter",
+  "AirDrop.label.keplr":"Keplr Wallet",
+  "AirDrop.label.LikerId":"Liker ID",
+  "AirDrop.label.missions":"Missions",
+  "AirDrop.label.myProgress":"My Progress",
+  "AirDrop.label.noWallet":"Continue without wallet",
+  "AirDrop.label.Osmosis":"Osmosis",
+  "AirDrop.mission.iscn":"We choose to go to the Moon in this decade and do the other things",
+  "AirDrop.mission.keplr":"We choose to go to the Moon",
+  "AirDrop.mission.liquidity":"Because that challenge is one that we are willing to accept",
+  "AirDrop.mission.stake":"Not because they are easy, but because they are hard",
+  "AirDrop.mission.vote":"Because that goal will measure the best of our energies ans skills",
+  "AirDrop.placeholder.email":"email address",
   "AppHeader.login.button": "Keplr",
   "AppHeader.tabBar.button.depub": "#DePub",
   "AppHeader.tabBar.button.publishing": "Your Publishing",
-  "AirDrop.button":"Airdrop",
-  "AirDrop.button.enter":"Enter",
-  "AirDrop.button.notify":"Get notify",
-  "AirDrop.button.claim":"Claim airdrop",
-  "AirDrop.content.prelaunch":"LikeCoin airdrop are on vote. The quick, brown fox jumps over a lazy dog. DJs flock by when MTV ax quiz prog. Junk MTV quiz",
-  "AirDrop.label.follow.LikeCoin":"Follow LikeCoin",
-  "AirDrop.label.follow.Twitter":"Follow us on Twitter",
-  "AirDrop.label.and":"and",
-  "AirDrop.label.$like":"$LIKE",
-  "AirDrop.label.cosmos":"Cosmos",
-  "AirDrop.label.decay":"Current Decay Factor",
-  "AirDrop.label.missions":"Missions",
-  "AirDrop.mission.keplr":"We choose to go to the Moon",
-  "AirDrop.mission.iscn":"We choose to go to the Moon in this decade and do the other things",
-  "AirDrop.mission.stake":"Not because they are easy, but because they are hard",
-  "AirDrop.mission.vote":"Because that goal will measure the best of our energies ans skills",
-  "AirDrop.mission.liquidity":"Because that challenge is one that we are willing to accept",
-  "AirDrop.label.myProgress":"My Progress",
-  "AirDrop.label.decay.start":"Time to Airdrop Decay start",
-  "AirDrop.label.Osmosis":"Osmosis",
-  "AirDrop.label.amount.umclaimed":"Unclaimed Airdrop",
-  "AirDrop.label.amount.airdrop":"Airdrop Amount",
-  "AirDrop.label.connect.wallet":"Connect your wallet",
-  "AirDrop.label.connect.address":"or insert your address",
-  "AirDrop.label.keplr":"Keplr Wallet",
-  "AirDrop.label.LikerId":"Liker ID",
-  "AirDrop.label.noWallet":"Continue without wallet",
-  "AirDrop.placeholder.email":"email address",
-  "AirDrop.guide.connect.wallet":"Connect or insert below your address and verify how many $LIKE have been allocated to you inside the upcoming LikeCoin airdrop",
-  "AirDrop.guide.connect.keplr":"Connect cosmos/osmosis wallet with Keplr broswer extension",
-  "AirDrop.guide.connect.LikerId":"Connect to LikerID with the Liker Land app",
   "AppHeader.tabBar.button.stake": "Delegate and Vote",
   "error.not.connect.wallet": "Please connect to your wallet first",
   "general.loading": "Loading",
